# backend/src/api/routes/metrics.py
"""
Metrics calculation endpoints for RAG evaluation.
"""
from fastapi import APIRouter, UploadFile, File, HTTPException
from pydantic import BaseModel
from typing import List, Dict, Any, Optional
import logging

router = APIRouter()
logger = logging.getLogger(__name__)


# Placeholder - will be implemented in Phase 3 (User Story 1)
@router.post("/validate")
async def validate_csv(file: UploadFile = File(...)) -> Dict[str, Any]:
    """
    Validate CSV file structure.
    
    Args:
        file: Uploaded CSV file
        
    Returns:
        Validation result with column info and preview
    """
    # Implementation in Phase 3
    raise HTTPException(status_code=501, detail="Not implemented yet")


@router.post("/metrics/calculate")
async def calculate_metrics(data: Dict[str, Any]) -> Dict[str, Any]:
    """
    Calculate RAG metrics from uploaded data.
    
    Args:
        data: Dataset with queries, responses, and ground truth
        
    Returns:
        Calculated metrics (aggregate and per-query)
    """
    # Implementation in Phase 3
    raise HTTPException(status_code=501, detail="Not implemented yet")


@router.post("/export/pdf")
async def generate_pdf_report(data: Dict[str, Any]) -> Any:
    """
    Generate PDF report with metrics and visualizations.
    
    Args:
        data: Metrics data to include in report
        
    Returns:
        PDF file
    """
    # Implementation in Phase 5 (User Story 3)
    raise HTTPException(status_code=501, detail="Not implemented yet")

# Made with Bob